@@ -6,7 +6,7 @@
     "clean": "rimraf dist",
     "webpack": "cross-env NODE_ENV=production webpack -p --config webpack.config.prod.js",
     "assets": "copy static/* dist && copy favicon.ico dist && copy CNAME dist",
-    "build:static": "run-s clean webpack assets",
+    "build": "run-s clean webpack assets critical",
     "start": "node devServer.js",
     "lint:js": "eslint src --ext .js --ext .jsx",
     "lint": "run-s lint:*",
@@ -14,7 +14,9 @@
     "critical": "critical ./dist/index.html --inline --base ./dist -H ./dist/index.html > /dev/null",
     "surge": "surge -p dist -d literate-army.surge.sh",
     "gh-pages": "gh-pages -d dist",
-    "deploy": "run-s build:static critical gh-pages",
+    "serve": "serve ./dist",
+    "dryrun": "run-s build serve",
+    "deploy": "run-s build gh-pages",
     "precommit": "run-s lint test"
   },
   "repository": {
@@ -27,58 +29,56 @@
     "autoprefixer": "^6.5.2",
     "babel-core": "^6.3.15",
     "babel-eslint": "^7.1.1",
-    "babel-jest": "^17.0.2",
+    "babel-jest": "^18.0.0",
     "babel-loader": "^6.2.8",
     "babel-plugin-lodash": "^3.2.10",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.18.0",
     "babel-preset-latest": "^6.16.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-react-hmre": "^1.0.0",
+    "babel-preset-stage-2": "^6.18.0",
     "critical": "^0.8.0",
     "cross-env": "^3.1.3",
-    "css-loader": "^0.26.0",
-    "dotenv": "^2.0.0",
-    "eslint": "^3.11.0",
-    "eslint-config-airbnb": "^13.0.0",
+    "css-loader": "^0.26.1",
+    "dotenv": "^4.0.0",
+    "eslint": "^3.11.1",
+    "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-babel": "^4.0.0",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.7.1",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.13.3",
     "extract-text-webpack-plugin": "^2.0.0-beta.4",
     "gh-pages": "^0.12.0",
     "html-webpack-plugin": "^2.24.1",
-    "husky": "^0.11.9",
-    "jest": "^17.0.3",
-    "lodash-webpack-plugin": "^0.10.6",
-    "node-sass": "^3.13.0",
-    "npm-run-all": "^3.1.1",
-    "offline-plugin": "^3.4.2",
+    "husky": "^0.13.1",
+    "jest": "^18.1.0",
+    "lodash-webpack-plugin": "0.11.4",
+    "node-sass": "^4.5.0",
+    "npm-run-all": "4.1.1",
+    "offline-plugin": "^4.5.2",
     "postcss-loader": "^1.1.0",
     "react-test-renderer": "^15.4.1",
-    "redux-devtools": "^3.0.1",
-    "redux-devtools-dock-monitor": "^1.0.1",
-    "redux-devtools-log-monitor": "^1.1.1",
-    "redux-logger": "^2.7.4",
     "rimraf": "^2.4.3",
-    "sass-loader": "^4.0.2",
+    "sass-loader": "^5.0.1",
+    "serve": "6.4.1",
     "style-loader": "^0.13.0",
-    "webpack": "^2.1.0-beta.25",
-    "webpack-dev-middleware": "^1.8.4",
-    "webpack-hot-middleware": "^2.13.2"
+    "webpack": "^2.2.1",
+    "webpack-dev-middleware": "^1.10.0",
+    "webpack-hot-middleware": "^2.16.1"
   },
   "dependencies": {
     "copy": "^0.3.0",
-    "file-loader": "^0.9.0",
-    "preact": "^6.4.0",
-    "preact-compat": "^3.9.2",
+    "file-loader": "^0.10.0",
+    "preact": "^7.1.0",
+    "preact-compat": "^3.9.3",
     "rc-animate": "^2.3.1",
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
-    "react-redux": "^4.4.6",
+    "react-router": "^3.0.0",
     "redux": "^3.6.0",
-    "redux-thunk": "^2.1.0",
+    "svgo": "^0.7.2",
     "svgo-loader": "^1.1.2",
     "url-loader": "^0.5.7"
   },
