@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
+import { Router, hashHistory } from 'react-router';
 
 /**
  * Import the stylesheet you want used! Here we just reference
@@ -8,42 +8,41 @@ import { Provider } from 'react-redux';
  */
 import './styles/main.scss';
 
-/**
- * Both configureStore and Root are required conditionally.
- * See configureStore.js and Root.js for more details.
- */
-import configureStore from './store/configureStore';
+import pages from './constants/pages';
 import App from './containers/App';
-
-const store = configureStore();
-
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root')
-);
+import Links from './components/links';
 
 /**
  * Add support for Service Worker
  * Reload the page automatically when updated code is detected
  */
-if (process.env.NODE_ENV === 'production') {
-  const runtime = require('offline-plugin/runtime') // eslint-disable-line
+import './offline';
 
-  runtime.install({
-    onUpdateReady: () => {
-      console.log('SW Event:', 'onUpdateReady');
+const getLink = (linkProps) => () => (<Links {...linkProps} />);
 
-      // Tells to new SW to take control immediately
-      runtime.applyUpdate();
-    },
+const routes = {
+  path: '/',
+  component: App,
+  indexRoute: { component: getLink(pages.index) },
+  childRoutes: Object.keys(pages).map((path) => ({ path, component: getLink(pages[path]) }))
+};
 
-    onUpdated: () => {
-      console.log('SW Event:', 'onUpdated');
+ReactDOM.render(
+  <Router history={hashHistory} routes={routes} />,
+  document.getElementById('root')
+);
 
-      // Reload page to load the new version
-      window.location.reload();
-    }
+/**
+ * 1) Prevent full-page reloads when clicking the header or "start over" links
+ * 2) Capture the outbound action links for Analytics
+ * (They aren't rendered by React, so bind to the window)
+ */
+[...document.querySelectorAll('[data-restart]')].forEach((el) => {
+  el.addEventListener('click', (e) => {
+    e.preventDefault();
+    hashHistory.push('/');
+
+    const { type, tag } = pages.index.analytics;
+    window.ga('send', 'event', type, tag);
   });
-}
+});
