@@ -54,10 +54,10 @@ module.exports = {
   ],
   resolve: {
     extensions: ['.js', '.jsx', '.json'],
-    alias:      {
-      'react':     'preact-compat',
-      'react-dom': 'preact-compat'
-    }
+    // alias:      {
+    //   'react':     'preact-compat',
+    //   'react-dom': 'preact-compat'
+    // }
   },
   module:  {
     loaders: [
